corrections with current version

+++ src/main/java/com/epam/lowcost/service/implementations/FlightServiceImpl.java
@@ -5,6 +5,7 @@ import com.epam.lowcost.DAO.interfaces.TicketDAO;
 import com.epam.lowcost.model.Flight;
 import com.epam.lowcost.service.interfaces.FlightService;
 import com.epam.lowcost.service.interfaces.PlaneService;
+import com.epam.lowcost.service.interfaces.TicketService;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -12,12 +13,15 @@ import java.util.List;
 public class FlightServiceImpl implements FlightService {
     private FlightDAO flightDAO;
     private PlaneService planeService;
-    private TicketDAO ticketDAO;
+    private TicketService ticketService;
 
-    public FlightServiceImpl(FlightDAO flightDAO, PlaneService planeService, TicketDAO ticketDAO) {
+    public FlightServiceImpl(FlightDAO flightDAO, PlaneService planeService) {
         this.flightDAO = flightDAO;
         this.planeService = planeService;
-        this.ticketDAO = ticketDAO;
+    }
+
+    public void setTicketService(TicketService ticketService){
+        this.ticketService = ticketService;
     }
 
     @Override
@@ -43,9 +47,10 @@ public class FlightServiceImpl implements FlightService {
 
     @Override
     public Flight deleteFlight(Long id) {
-        ticketDAO.deleteTicketsByFlightId(id);
-        return flightDAO.deleteFlight(id);
-
+        if (ticketService.deleteTicketsByFlightId(id)) {
+            return flightDAO.deleteFlight(id);
+        }
+        return null;
     }
 
     @Override
